change tarball extraction to use any top level directory
Rather than trying to use the exact filename of the tarball or having
special cases for both cperl and blead, just check for a top level
directory in the tarball and use that as the extracted root.

This removes the special cases, and allows the code to work with
other sources where the tarball name may not match its contents.