Use the mtime of the cached blead.tar.gz for "as"
While it would be nicer to use the time of the last commit, I don't have
guaranteed access to an https client, since HTTP::Tinyish is fat-packed
into perl-build, which means I can't query the Github API for it.

So instead, reset the mtime after we've finished installing so that next
time it will say we've already installed it.